+++ service/controller/resources/vaultcrt/create.go
@@ -6,6 +6,8 @@ import (
 	"github.com/giantswarm/microerror"
 	apiv1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	apiv1alpha2 "sigs.k8s.io/cluster-api/api/v1alpha2"
 
 	"github.com/giantswarm/cert-operator/service/controller/key"
 )
@@ -21,6 +23,18 @@ func (r *Resource) ApplyCreateChange(ctx context.Context, obj, createChange inte
 		if err != nil {
 			return microerror.Mask(err)
 		}
+
+		r.logger.LogCtx(ctx, "level", "debug", "message", "finding cluster resource")
+		cluster := &apiv1alpha2.Cluster{}
+		if err := r.ctrlClient.Get(ctx, types.NamespacedName{
+			Namespace: customObject.Namespace,
+			Name:      key.ClusterID(customObject)},
+			cluster); err != nil {
+			return microerror.Maskf(notFoundError, "Could not find cluster %s in namespace %s.",
+				customObject.Namespace,
+				key.ClusterID(customObject))
+		}
+
 		r.logger.LogCtx(ctx, "level", "debug", "message", "creating the secret in the Kubernetes API")
 
 		_, err = r.k8sClient.CoreV1().Secrets(customObject.GetNamespace()).Create(ctx, secretToCreate, metav1.CreateOptions{})